Refactor back-end architecture
